requirements.txt
astroid==2.2.5
autopep8==1.4.4
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-admin-tools==0.8.1
django-allauth==0.39.1
django-appconf==1.0.3
django-controlcenter==0.2.9
django-cors-headers==3.0.2
django-countries==5.4
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-extensions==2.2.1
django-filer==1.5.0
django-fsm==2.6.1
django-fsm-admin==1.2.4
django-import-export==1.2.0
django-js-asset==1.2.2
django-localflavor==2.2
django-material-admin==1.4.2
django-mptt==0.10.0
django-phone-field==1.8.0
django-pkgconf==0.3.0
django-polymorphic==2.0.3
django-rest-auth==0.9.5
django-select2==7.1.1
django-smart-selects==1.5.4
django-suit==0.2.26
djangorestframework==3.9.4
easy-thumbnails==2.6
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.3
pep8==1.7.1
Pillow==6.1.0
pycodestyle==2.5.0
pydot==1.4.1
pylint==2.3.1
pyparsing==2.4.2
python-stdnum==1.11
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
typed-ast==1.4.0
Unidecode==1.0.23
urllib3==1.25.3
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
